+++ addons/mail/static/src/discuss/core/common/discuss_channel_model.js
@@ -488,6 +488,35 @@ export class DiscussChannel extends Record {
         }
     }
 
+    async leaveChannel() {
+        if (this.channel_type === "channel" && this.create_uid?.eq(this.store.self_user)) {
+            await this.askLeaveConfirmation(
+                _t("You are the administrator of this channel. Are you sure you want to leave?")
+            );
+        }
+        if (this.channel_type === "group") {
+            await this.askLeaveConfirmation(
+                _t(
+                    "You are about to leave this group conversation and will no longer have access to it unless you are invited again. Are you sure you want to continue?"
+                )
+            );
+        }
+        await this.leaveChannelProcess();
+    }
+
+    async leaveChannelProcess() {
+        await this.closeChatWindow();
+        if (this.exists()) {
+            await this.leaveChannelRpc();
+        }
+    }
+
+    async leaveChannelRpc() {
+        await this.store.env.services.orm.silent.call("discuss.channel", "action_unfollow", [
+            this.id,
+        ]);
+    }
+
     async markAsFetched() {
         await this.store.env.services.orm.silent.call("discuss.channel", "channel_fetched", [
             [this.id],
